+++ packages/x-components/src/x-installer/api/api.types.ts
@@ -21,6 +21,16 @@ export interface XAPI {
    * getting the previously stored result information.
    */
   addProductToCart(productId?: string): void;
+
+  /**
+   * To get the snippet config object.
+   *
+   * @returns The {@link NormalisedSnippetConfig | snippetConfig} object.
+   *
+   * @public
+   */
+  getSnippetConfig(): SnippetConfig;
+
   /**
    * To set the {@link @empathyco/x-bus#XBus | bus} to the API. This bus will be used to emit
    * the necessary events.
@@ -47,6 +57,15 @@ export interface XAPI {
    * @public
    */
   setSnippetConfig(config: Partial<SnippetConfig>): void;
+
+  /**
+   * To set or update any property of the snippet config getter.
+   *
+   * @param snippetConfigGetter - A function that returns the snippet config object.
+   *
+   * @internal
+   */
+  setSnippetConfigGetter(snippetConfigGetter: () => NormalisedSnippetConfig): void;
 
   /**
    * To set the snippet config callback.